+++ frontend/src/backend/web/clipboard.ts
@@ -10,3 +10,16 @@ export const getClipboardText = async () => {
     return s === null ? Promise.reject(new Error("User cancel")) : s;
   }
 };
+
+export const setClipboardText = async (text: string) => {
+  try {
+    await window.navigator.clipboard.writeText(text);
+  } catch (e) {
+    const note = e instanceof Error ? ` (${e.message})` : "";
+    console.error("navigator.clipboard.writeText():", e);
+    window.prompt(
+      `No access or support for clipboard${note}\n\nCopy this:`,
+      text,
+    );
+  }
+};